+++ src/main/java/project/trendpick_pro/global/tosspayment/controller/PaymentController.java
@@ -4,7 +4,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -14,7 +13,6 @@ import project.trendpick_pro.domain.common.base.rq.Rq;
 import project.trendpick_pro.domain.member.entity.Member;
 import project.trendpick_pro.domain.notification.service.NotificationService;
 import project.trendpick_pro.domain.orders.entity.Order;
-import project.trendpick_pro.domain.orders.entity.OrderStatus;
 import project.trendpick_pro.domain.orders.service.OrderService;
 import project.trendpick_pro.global.rsData.RsData;
 import project.trendpick_pro.global.tosspayment.dto.PaymentResultResponse;
